Deleted Active Node Syncing
Depreciated Active Node Syncing - it is not required as the server just will not respond when syncing known nodes! - the only thing that may have to change is the verification while() loop at the start -> the numOfActiveNodes int may have to be adjusted dynamically to the number of nodes that are pinging back (aka. communicating)