Added SyncTransactionFee Implementation
Added an implementation for every function in the header except for sync() - this will come soon. The way this framework works is that there are 2 threads (1 for expiry time and 1 for updating fee's). A program can call getTransactionFee() with the relevant inputs to get the transaction fee's. This framework does not require any node communication (only pings nodes to detect if they are active or not).